added in checking to the IE script

+++ webpagetest/powershell/Set-DisableIESecurity.ps1
@@ -1,9 +1,14 @@
-function Disable-IEESC
-{
 $AdminKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A7-37EF-4b3f-8CFC-4F3A74704073}"
 $UserKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A8-37EF-4b3f-8CFC-4F3A74704073}"
-Set-ItemProperty -Path $AdminKey -Name "IsInstalled" -Value 0
-Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value 0
-Stop-Process -Name Explorer
+
+$CurrentVal = Get-ItemProperty -Path $AdminKey -Name "IsInstalled"
+
+if ($CurrentVal -ne 0) {
+  Disable-IEESC
+  Set-ItemProperty -Path $AdminKey -Name "IsInstalled" -Value 0
+  Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value 0
+  Stop-Process -Name Explorer
+  Write-Output "IE ESC Disabled."
+} Else {
+  Write-Output "IE ESC Already Disabled."
 }
-Disable-IEESC
